Refactor: 게시판 페이지 하나로 통일
