+++ src/components/global/NavBar/index.tsx
@@ -1,9 +1,8 @@
-import styled from 'styled-components';
-import { NavLink } from 'react-router-dom';
-
-import COLOR from '../../../constants/color';
-
-import { LogoIcon } from '../../Icons/LogoIcons';
+import styled from "styled-components";
+import { NavLink } from "react-router-dom";
+import COLOR from "../../../constants/color";
+import { LogoIcon } from "../../Icons/LogoIcons";
+import { useBoardCategoryList } from "../../../hooks/board/useBoardCategoryList";
 import {
   DoubleLeftIcon,
   HomeIcon,
@@ -12,9 +11,12 @@ import {
   CalIcon,
   MedalIcon,
   EyeIcon,
-} from '../../Icons/BasicIcons';
+} from "../../Icons/BasicIcons";
+import BoardCategory from "../../board/BoardCategory";
 
 const NavBar = () => {
+  const { boardCategoryList } = useBoardCategoryList();
+
   return (
     <Wrapper>
       <Top>
@@ -22,7 +24,7 @@ const NavBar = () => {
         <DoubleLeftIcon />
       </Top>
       <Links>
-        <Link to="/home">
+        <Link to="/">
           <HomeIcon />홈
         </Link>
         <BaseLink>
@@ -37,13 +39,9 @@ const NavBar = () => {
           <Blank />
           역할 관리
         </Link>
-        <Link to="/setting/character">
+        <Link to="/setting/board">
           <Blank />
-          역할 관리
-        </Link>
-        <Link to="/setting/character">
-          <Blank />
-          역할 관리
+          게시판 관리
         </Link>
         <BaseLink>
           <CheckRoundIcon />
@@ -53,25 +51,11 @@ const NavBar = () => {
           <BoardIcon />
           게시판
         </BaseLink>
-        <Link to="/board/announcement">
+        <Link to="/board">
           <Blank />
-          공지사항
-        </Link>
-        <Link to="/board/question">
-          <Blank />
-          질문게시판
-        </Link>
-        <Link to="/board/free">
-          <Blank />
-          자유게시판
-        </Link>
-        <Link to="/board/bug">
-          <Blank />
-          버그가 있어요!
-        </Link>
-        <Link to="/board/changelog">
-          <Blank />
-          체인지 로그
+          {boardCategoryList && (
+            <BoardCategory boardCategoryList={boardCategoryList} />
+          )}
         </Link>
         <Link to="/schedule">
           <CalIcon />
