+++ src/routes/Board/BoardLayout/index.tsx
@@ -1,8 +1,8 @@
 import { Outlet } from "react-router-dom";
-import FONT from "./../../../../constants/fonts";
+import FONT from "../../../constants/fonts";
 import { Container, Title } from "./style";
 
-const AnnouncementLayout = () => {
+const BoardLayout = () => {
   return (
     <Container>
       <Title style={FONT.HEADING}>공지사항</Title>
@@ -11,4 +11,4 @@ const AnnouncementLayout = () => {
   );
 };
 
-export default AnnouncementLayout;
+export default BoardLayout;
