+++ src/routes/Board/BoardPage/index.tsx
@@ -3,12 +3,9 @@ import BoardCreateButton from "../../../components/board/BoardCreateButton";
 import BoardList from "../../../components/board/BoardList";
 import { useBoardList } from "../../../hooks/board/useBoardList";
 import { BoardButtonContainer, BoardContainer } from "./style";
-import { useBoardCategoryCreate } from "../../../hooks/category/useBoardCategoryCreate";
 
-const Announcement = () => {
+const BoardPage = () => {
   const { boardList } = useBoardList(1, 1, 5);
-
-  const createMutation = useBoardCategoryCreate("공지사항");
 
   return (
     <>
@@ -18,10 +15,8 @@ const Announcement = () => {
       <BoardButtonContainer>
         <BoardCreateButton />
       </BoardButtonContainer>
-
-      <button onClick={() => createMutation.mutate()}>공지사항 생성</button>
     </>
   );
 };
 
-export default Announcement;
+export default BoardPage;
